Updated project dependencies to Developist.Core.Persistence v5. Restructured project organization, including namespace refactoring. Renamed the project to Developist.Core.Persistence.Extensions.
