+++ Developist.Core.Persistence.Extensions.sln
@@ -3,13 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.2.32630.192
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Developist.Extensions.Persistence", "Developist.Extensions.Persistence\Developist.Extensions.Persistence.csproj", "{0E1AFD5E-6CCD-44FD-9EF2-367054C8435F}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{5B5B095D-C0FF-4C8D-80E9-D791FD8FD157}"
-	ProjectSection(SolutionItems) = preProject
-		Directory.Build.props = Directory.Build.props
-		Version.props = Version.props
-	EndProjectSection
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Developist.Core.Persistence.Extensions", "src\Developist.Core.Persistence.Extensions\Developist.Core.Persistence.Extensions.csproj", "{0E1AFD5E-6CCD-44FD-9EF2-367054C8435F}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
