--- Directory.Build.props
@@ -1,26 +1,0 @@
-<Project>
-  <PropertyGroup>
-    <Product>Developist.Extensions.Persistence</Product>
-    <Authors>Jim Atas</Authors>
-    <Copyright>Copyright (c) $([System.DateTime]::Now.Year) $(Authors)</Copyright>
-  </PropertyGroup>
-  <Import Project="Version.props"/>
-  <PropertyGroup>
-    <VersionPrefix>$(Major).$(Minor).$(Patch)</VersionPrefix>
-    <AssemblyVersion>$(Major).$(Minor).0.0</AssemblyVersion>
-    <FileVersion>$(Major).$(Minor).$(Build).$(Revision)</FileVersion>
-  </PropertyGroup>
-  <PropertyGroup>
-    <GenerateDocumentationFile>true</GenerateDocumentationFile>
-    <NoWarn>1591</NoWarn>
-  </PropertyGroup>
-  <PropertyGroup>
-    <Description>Provides useful extensions to the Developist.Core.Persistence package.</Description>
-    <PackageTags>entityframeworkcore;efcore;dbcontext;repository;unitofwork</PackageTags>
-    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-    <PackageLicenseExpression>MIT</PackageLicenseExpression>
-    <PackageProjectUrl>https://github.com/jimatas/Developist.Extensions.Persistence</PackageProjectUrl>
-    <RepositoryUrl>https://github.com/jimatas/Developist.Extensions.Persistence</RepositoryUrl>
-    <RepositoryType>git</RepositoryType>
-  </PropertyGroup>
-</Project>
