--- Version.props
@@ -1,9 +1,0 @@
-<Project>
-  <PropertyGroup>
-    <Major>1</Major>
-    <Minor>1</Minor>
-    <Patch>0</Patch>
-    <Build>$([System.Math]::Floor($([System.DateTime]::Now.Subtract($([System.DateTime]::Parse("01-01-2000"))).TotalDays)))</Build>
-    <Revision>$([System.Math]::Floor($([MSBuild]::Divide($([System.DateTime]::Now.Subtract($([System.DateTime]::Today)).TotalSeconds), 2.0))))</Revision>
-  </PropertyGroup>
-</Project>
